Created FramedLZ4CompressorOutputStreamTest and moved framed tests out of block tests.
